Move AbstractAuthenticatingServerCallListener to own file
