+++ grpc-server-spring-boot-autoconfigure/src/main/java/net/devh/boot/grpc/server/security/interceptors/AuthenticatingServerInterceptor.java
@@ -21,16 +21,14 @@ import org.springframework.security.core.Authentication;
 
 import io.grpc.Context;
 import io.grpc.Contexts;
-import io.grpc.ForwardingServerCallListener.SimpleForwardingServerCallListener;
 import io.grpc.Metadata;
 import io.grpc.ServerCall;
 import io.grpc.ServerCall.Listener;
 import io.grpc.ServerCallHandler;
 import io.grpc.ServerInterceptor;
-import lombok.extern.slf4j.Slf4j;
 
 /**
- * A server interceptor that used to authenticate the client request.
+ * Marker-Interface: A server interceptor that used to authenticate the client request.
  *
  * <p>
  * <b>Note:</b> Implementations must be thread safe and return a thread safe {@link Listener}. Do <b>NOT</b> store the
@@ -40,6 +38,8 @@ import lombok.extern.slf4j.Slf4j;
  * </p>
  *
  * @author Daniel Theuke (daniel.theuke@heuboe.de)
+ * @see AbstractAuthenticatingServerCallListener
+ * @see Contexts#interceptCall(Context, ServerCall, Metadata, ServerCallHandler)
  */
 public interface AuthenticatingServerInterceptor extends ServerInterceptor {
 
@@ -47,131 +47,5 @@ public interface AuthenticatingServerInterceptor extends ServerInterceptor {
      * The context key that can be used to retrieve the associated {@link Authentication}.
      */
     public static final Context.Key<Authentication> AUTHENTICATION_CONTEXT_KEY = Context.key("authentication");
-
-    /**
-     * A call listener that will set the authentication context before each invocation and clear it afterwards. Use and
-     * extend this class if you want to setup non-grpc authentication contexts.
-     *
-     * <p>
-     * <b>Note:</b> If you only want to setup the grpc-context and nothing else, then you can use
-     * {@link Contexts#interceptCall(Context, ServerCall, Metadata, ServerCallHandler)} instead.
-     *
-     * @param <ReqT> The type of the request.
-     */
-    @Slf4j
-    abstract class AbstractAuthenticatingServerCallListener<ReqT> extends SimpleForwardingServerCallListener<ReqT> {
-
-        private final Context context;
-
-        /**
-         * Creates a new AbstractAuthenticatingServerCallListener which will attach the given security context before
-         * delegating to the given listener.
-         *
-         * @param delegate The listener to delegate to.
-         * @param context The context to attach.
-         */
-        protected AbstractAuthenticatingServerCallListener(final Listener<ReqT> delegate, final Context context) {
-            super(delegate);
-            this.context = context;
-        }
-
-        /**
-         * Gets the {@link Context} associated with the call.
-         *
-         * @return The context of the current call.
-         */
-        protected final Context context() {
-            return this.context;
-        }
-
-        /**
-         * Attaches the authentication context before the actual call.
-         *
-         * <p>
-         * This method is called after the grpc context is attached.
-         * </p>
-         */
-        protected abstract void attachAuthenticationContext();
-
-        /**
-         * Detaches the authentication context after the actual call.
-         *
-         * <p>
-         * This method is called before the grpc context is detached.
-         * </p>
-         */
-        protected abstract void detachAuthenticationContext();
-
-        @Override
-        public void onMessage(final ReqT message) {
-            final Context previous = this.context.attach();
-            try {
-                attachAuthenticationContext();
-                log.debug("onMessage - Authentication set");
-                super.onMessage(message);
-            } finally {
-                detachAuthenticationContext();
-                this.context.detach(previous);
-                log.debug("onMessage - Authentication cleared");
-            }
-        }
-
-        @Override
-        public void onHalfClose() {
-            final Context previous = this.context.attach();
-            try {
-                attachAuthenticationContext();
-                log.debug("onHalfClose - Authentication set");
-                super.onHalfClose();
-            } finally {
-                detachAuthenticationContext();
-                this.context.detach(previous);
-                log.debug("onHalfClose - Authentication cleared");
-            }
-        }
-
-        @Override
-        public void onCancel() {
-            final Context previous = this.context.attach();
-            try {
-                attachAuthenticationContext();
-                log.debug("onCancel - Authentication set");
-                super.onCancel();
-            } finally {
-                detachAuthenticationContext();
-                log.debug("onCancel - Authentication cleared");
-                this.context.detach(previous);
-            }
-        }
-
-        @Override
-        public void onComplete() {
-            final Context previous = this.context.attach();
-            try {
-                attachAuthenticationContext();
-                log.debug("onComplete - Authentication set");
-                super.onComplete();
-            } finally {
-                detachAuthenticationContext();
-                log.debug("onComplete - Authentication cleared");
-                this.context.detach(previous);
-            }
-        }
-
-        @Override
-        public void onReady() {
-            final Context previous = this.context.attach();
-            try {
-                attachAuthenticationContext();
-                log.debug("onReady - Authentication set");
-                super.onReady();
-            } finally {
-                detachAuthenticationContext();
-                log.debug("onReady - Authentication cleared");
-                this.context.detach(previous);
-            }
-        }
-
-    }
 
 }
